separate unpacking logic
